requirements.txt
fastapi
uvicorn
scikit-learn
pandas
joblib
kaggle
python-dotenv
